Removed circular dependency to finish tracer implementation
