+++ sympl/_core/dataarray.py
@@ -1,6 +1,6 @@
-from .units import data_array_to_units as to_units_function
 import xarray as xr
 from pint.errors import DimensionalityError
+from .units import data_array_to_units as to_units_function
 
 
 class DataArray(xr.DataArray):
